+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## Bug fixes 🐞
 * [compose] Fixed a warning that using UI composable where Mapbox Map composable is expected.
+* Fix a crash occuring when clicking on the "Telemetry settings" option in the attribution dialog when not using the `AppCompat` theme.
 
 # 11.2.0-beta.1 February 01, 2024
 ## Features ✨ and improvements 🏁